requirements.txt
flask
flask-cors
sqlalchemy
python-dotenv
psycopg2-binary
PyJWT
requests
cloudinary
